Render nginx templates into sites-available
Render nginx templates into sites-available and then symlink them.
Instead of rendering them directly to sites-enabled.

It's a good practice and makes it easier for disabling or substituting temporarily a site config.

Currently, the sites-enabled file is a regular file. If we just try to replace it with a symlink, it fails. To avoid upgrade conflicts, i add 2 tasks to remove the file if it's regular instead of a symlink. We could just delete it always before symlinking, but then we would generate unnecessary activity and "changed" noise in the output.

Tested with a running instance.